server/ai/requirements.txt
numpy
scipy
pandas
Pillow
cntk